refactor(broker): inject cluster services via system context
Refactors to allow the cluster services to be created outside of the
broker, and injected directly, though not yet started. This reduces the
cluster step to just starting the cluster.

To simplify testing, temporary factories are added, mostly because many
tests expect to be able to modify the cluster services by modifying the
`BrokerCfg`. Most of `ClusterConfigFactory` should be transferable to
the `dist` module once this is not true anymore.