fix: run now during inProcessing
Previously we avoid to run scheduled tasks if currently a processing was going on.
This is not really necessary and have potential negative impacts of tasks,
which are then delayed unnecessarily. For example it might happen that tasks are
delayed until no new record was committed on the log (worst case), which might
introduce a huge delay to timers and other things which have been scheduled.

All necessary task we want to run during processing AFTER a commit are executed
as PostCommitTasks, which means we no longer need thus gurantee