refactor(broker): prevent using an unitialized BrokerCfg
Attempts to limit usages of uninitialized `BrokerCfg` objects by
removing the public init call and instead forcing callers to use the
constructor. The default constructor is kept for testing purposes, as
many places in the tests still use uninitialized configuration objects.
We could think about removing these in the future, or adding a warning.