feat(broker): add partition restore service to restore backup of a partition
Restores data of a partition from a given backup to the configured data directory. After downloading the backup, it trucates the journal to the checkpointPosition. This is necessary to ensure that the backups of all partitions form a consistent checkpoint. After restore is done, the state of partition is consistent. A broker can successfully process from this state.