test(it): accept modify command
This tests that a process instance can be modified from the java client,
using multiple activation instructions with scoped variables and a
terminate instruction. This test only needs to test the communication
between the client and the engine. It does not need to verify that the
engine actually modified the instance. Checking that it wasn't rejected
is enough as there are other tests that show case rejections.