feat(broker): do no install services when backup feature flag is disabled
BackupStore and CheckpointProcessor are not installed. CheckpointRecord is not injected into the StreamProcessor. This is required because, checkpoint records written when the feature was enabled might still be in the log. When the feature is disabled, those records should not be processed or replayed.
BackupManager is still as a NoopBackupManager.