test(it): expand create instance anywhere
The QA test should verify the integration between the client and the
engine, making sure that all the instructions are applied.

We can verify this using the RecordingExporter.

Note that the intent of this test is not to verify the entire
CreateProcessInstance with StartInstructions behavior, but it's
important to verify that the instructions are used.