feat(util): add VisibleForTesting annotation
Adds a new compile-time only annotation indicating when certain types,
fields, etc., have stronger visibility than necessary for testing. While
this is not the preferred solution, it's sometimes useful for unit
tests with minimal downsides.

Note that this should never be used for public visibility items, as
that's an even stronger indication that something is broken.