feat(protocol): add EXCEEDED_BATCH rejection type
Adds a new rejection type which explicitly states that the rejection was due to exceeding the batch size.
This is being introduced to prevent rejection-loops. These loops could occur when a rejection is being written that's exceeding the batch. To make sure we are able to write these rejections we have to omit the rejection reason. In order to still provide some context on why a command was rejected this rejection type is set.