refactor(atomix): calculate partition members when building the partition group
Previously it was calculated when starting the partition. But all necessary info to determin the partition members are already available when building the group. For restore app, we have to determine the partition members, but there is no need to start the partitions. So it is moved from `join` to the constructor.