feat(engine): prevent exception loop
Handling an unexpected error could result in an exception loop. The reason for this is that writing the rejection with a large rejection reason results in an ExceededBatchRecordSizeException, as the rejection reason contains a String representation of the large record.
By leaving the rejection reason empty we can guarantee that the rejection event will not become too large. In order to be able to identify the reason for the rejection the new EXCEEDED_BATCH rejection type will be used.