fix(engine): remove exceeded batch exception handling
This is now handled in a more generic way. The downside of removing this here is that the response to the user is a lot more abstract and could be unclear.

The advantage is that this way of dealing with the problem was not fool-proof. A command with a specific size + the custom exception message could still be too large. This would create another exception-loop.