refactor: make ScheduleServiceImpl an own actor
The actor between StreamProcessor and ProcessingScheduleServiceImpl is
no longer shared. ProcessingScheduleServiceImpl is scheduled as own
actor.

ProcessingScheduleServiceImpl is created before the RecordProcessors are
initialized. This allows to initialize all processors already correctly
on init.

The ProcessingScheduleServiceImpl actor is submitted to the scheduler
AFTER replay is done. Only after that scheduled tasks are accepted.

The ProcessingScheduleServiceImpl gets on creation some suppliers which
are used during the actors lifetime. For example an LogStreamWriter
supplier, which is used to create an own writer AFTER the actor is
scheduled. This reduces the consumed resources on followers, if the
sercice is never scheduled we don't need to create a writer.

When the StreamProcessor is in closing phase it will close the schedule
service as well.