test(engine): fix flaky JobMetricsTest
This test activates a job with a timeout of 10 minutes. After activating the engine time is increased by 10 minutes. The test will now verify that the job has run into a timeout.
The assertion we do has a maximum waiting time of 5 seconds. The JobTimeoutTrigger is scheduled once every 30 seconds. This results in a relatively high margin of error. The expectation would be that the test fails 1 in 6 times.
This is not the case. Most of the time the test succeeds. My assumption is that the increase in engine time will cause the JobTimeoutTrigger to be executed as the increase > 30 seconds.
The activating of the Job takes a few milliseconds. This means that in reality this job does not time out after 10 minutes exactly, but after 10 minutes + a few milliseconds. If the JobTimeoutTrigger triggers immediately when the engine time is increased this could mean that the job still has a few milliseconds before it's supposed to time out. These few milliseconds are the margin of error. Since the next trigger will be in 30 seconds, the job won't time out within the 5 seconds we try to assert.

By increasing the engine time with 50ms more, we know for sure that the Job has passed it's timeout time and the trigger will write the TIMEOUT command.