feat(dist): add get backup status API
Adds a new actuator endpoint to get the status of a backup. This
endpoint is simply `GET /{id}` of the backup under whichever route the
actuator is set up.

Currently we do not do any response mapping and use the gateway types
directly. While this is not a good approach long term (as it exposes
internal types), this is fine for now as we will later use the OpenAPI
spec which will generate models for us, and then we will do the mapping.
Doing the mapping to our custom types now would end up simply creating
busy work we would throw out later.