feat(qa/util): add backup status API to BackupActuator
Adds mapping to get the backup status to the `BackupActuator`. Again we
reuse the same internal types, but later this whole client will be
mostly generated by the OpenAPI generator.

One thing to note, we need to add the `Jdk8Module` to our
`JacksonDecoder` - this is done on the server side automatically
by Spring, but here we have to do it manually to decode `Optional`
types.