refactor: replace usage of additional Actor#submit
Iterate over the ProcessingScheduleServiceImpl and remove the
StreamProcessorContext from the constructor.

The service gets all necessary dependencies injected on the constructor
(except the actor),
mostly as suppliers. This allows to lazy load dependencies and reduce
dependencies for tests.

The service gets an open method to initialize the service with the actor
control and create an own writer. This allows to no longer share the
writer with the StreamProcessor.

Due to later initializing of the writers we reduce resources on
followers, if the
sercice is never scheduled we don't need to create a writer.

When the StreamProcessor is in closing phase it will close the schedule
service as well.