feat(engine): the process instance handles terminate end events
The process instance reached a terminate end event in the process that terminated all element instances of the process. After all element instances are terminated, we complete the process instance.

We check the current state of the process instance to distingulish the behavior of a regular cancallation of the process instance, or an interruption from a parent process instance.