test(engine): verify the cancellation behavior of terminate end events
If the process instance reaches a terminate end event in a subprocess then we should terminate all active element instances inside the subprocess. After the element instances are terminated, we should complete the subprocess and take the outgoing sequence flow. Other element instances outside the subprocess should not be terminated.