build(.github): configure Maven for CI timeouts
Sets sane defaults to configure Maven to be more resilient when it comes
to timeouts when downloading dependencies. This is a problem which
primarily affects CI, and notably GitHub Actions runners.

To set the options globally, a new step is added in `setup-zeebe` iff
maven is being set up there, setting the environment variable for
further workflow steps.

To make sure this is used everywhere, all usages of setup-maven and
setup-java must thus be replaced with setup-zeebe itself. To simplify
this, we now always create a default `settings.xml` which is set up for
deployment (but ignored when not deploying), and add a flag to use the
internal Camunda Nexus cache for self hosted runners.