fix(raft): do not fail consistency check if log is empty
The check for consistency is moved to a util class, so that it can be tested using unit test. Otherwise, the only way to test is by starting a cluster of 3 nodes. This is too cumbersome, as you can see in the RaftConsistencyCheckTest.

(cherry picked from commit 32e67ec70e09716ec71083e7c0e5efa3ee7b668a)