test(engine): speed up the inclusive gateway test
The test case to verify activated tasks took 15 seconds to complete. In the test, we limited the record stream by the completion of the process instance. Since the process instance were not completed, the test waited 5 seconds until reaching the timeout. This was done 3 times.

Changing the service tasks in the process to manual tasks. Manual tasks has a pass-through and are immediately completed. As a result, we can limit the record stream because the process instance is immediately completed. The service tasks were not semantical important for the test case.