feat(build-tools): number all test output per run
To ensure that we preserve test output of flaky tests - that is, the
output of a test even if it's re-run - we use a custom implementation of
the `ConsoleOutputReporter`. This implementation is a simple delegating
implementation, except for to return a custom listener which will ensure
that all run outputs are numbered (e.g. 1, 2, 3, etc.) for each test.
The custom listener is also a delegating one, and simply hooks into the
lifecycle of the canonical implementation. After writing the output
(e.g. `MyTest-output.txt`), the listener will compute the run count from
the file system (to avoid caching an inordinate amount of test files)
and move to the file to its proper numbered name, e.g.
`MyTest-1-output.txt` or `MyTest-2-output.txt`).

As the canonical implementation is not meant to be extended, we opted
for this delegating pattern which is a little hacky (as we "predict" the
file name of the report via reflection), but it's a good compromise in
terms of keeping up with Surefire improvements and maintaining our
complete implementation.