ci(test): balanced runOrder
This sorts all test by their runtime if statistics are available. When forks are used this results in a more balanced distribution of tests across the forks. Although not fully optimal it already increases the chances for reduced total runtimes as forks pull tests out of the ordered test queue.
More improvements, for as even as possible fork runtimes, are only possible once https://issues.apache.org/jira/browse/SUREFIRE-1387 is implemented.