test: add verification timeout
The test was not reliable since the ordering of the command response writer wock was not the same as in the code, this means it could happen that the writer was not yet called on all methods. It was easier to reproduce and also to fix.

(cherry picked from commit 20fa4cb258e0617d7b1e606dce5b457c132a0022)