feat(engine): generate terminate end event compatible execution steps part 1
The random execution tests don't know the concept of flow scopes. This makes it challenging to generate a correct execution path for terminate end events, as they terminate a specific flow scope. Processing should continue as normal once the flow scope has been terminated.

Whilst we don't have flow scopes, we do have execution path segments. These segments don't map 1 to 1 to flow scopes. However, since every flow scope starts a new segment we can use these segments to get the desired behavior.

Each segment must keep track whether is has reached a terminate end event. If this is the case that means that we don't expect any further execution steps. We can isolate this behavior in a single location, during the appending of one segment to another segment.
In order to differentiate between flow scopes a new append method has been added which takes the boolean `changesFlowScope` as a parameter. Blockbuilder where the flow scope changes (e.g. SubProcessBlockBuilder) can use this to indicate that even though a terminate end event has been reached. Execution steps after this specific segment still need to added to complete the process.

When a segment is appended to a different segment and the flow scope does not change we can use the segment that should be appended to identify whether new segment can still be added to the current segment. If passed segment has reached a terminate end event and the flow scope has not been changed it is guaranteed that the current segment is in the same flow scope has the previous segment and thus has also reached the terminate end event.