fix(engine): prevent NPE during termination
A command could contain multiple terminate instruction that result in the termination of the same element instance. For example, when we have a sub process and a nested element. If both this sub process and the nested element are part of the terminate instructions, the nested element will be terminated twice.

When we receive the command we verify that all instances exist. However, once the first terminate instruction terminates the element, the instance no longer exist for the second terminate instruction. When this occurs we should ignore the terminate instruction.

(cherry picked from commit 43b0a1d197e094df49411291e1c4647c24090d4b)