test: fix unfinished stubbing of command response writer
The engine is shared across multiple tests and not stopped in-between.
With this change we only set up mocking of the response writer once and
not for every test method. This ensures that mocking does not interfere
with engine processing from previous tests which would show up as
unfinished mocking exceptions.