fix: take snapshot if nothing was exported since last snapshot
When figuring out where to take the next snapshot, we determine the
snapshot position as the minimum of processing and exporter positions.

There was an edge case where a leader could process but no export.
In that case it'd use the exporter position as snapshot position and try
and find a log entry at that position.
If the log starts with the exporter position, for example
because the same broker previously received a snapshot and compacted the
log, no entry could be found which led to a failed snapshot.

We now try and use the latest snapshot's term and index if the log entry
could not be found. This ensures that new snapshots can be taken even if
nothing was exported since the last snapshot.

(cherry picked from commit ef090ab9b374e24deceafb25398fc23b37d2e01f)