feat(broker): reduce log level of received exporter position
Logging for every update when receiving exporter position was flooding the logs. To reduce the noise, lowered the log level to trace.

In addition, added a new log which will be logged periodically at every 60 seconds. This would be enough in most cases to monitor the exporter position. There are two logs because in case needed during debugging, the log level of the trace can be increased to debug dynamically, which will then logged for every message. The frequency of periodic logger cannot be increased dynamically.