test: close manually to avoid leaks in `ClusteringRuleExtension`
`ClusteringRule` uses rule chaining to ensure that gateway and client
resources will be closed when tests are finished. We wrapped the
`ClusteringRule` in `ClusteringRuleExtension` for use in JUnit5.
The extension calls `after` and `before` callbacks of the rule, but it
does not call `apply` which is responsible for making sure that inner
rules such as the `AutoCloseableRule` are also informed about the
test lifecycle.

This was already an issue for the recording exporter and that's why the
extension had to implement this part again and couldn't rely on
`ClusteringRule`. With this change, we let `ClusteringRule` handle
closeables directly, without relying on the `AutoCloseableRule`. This
ensures that the `after` callback from the extension will actually close
all resources created by `ClusteringRule`.