test: clarify that test just waits for snapshot
This test case looked a bit misleading because it was waiting for a
stable exported position in the latest snapshot. This was unnecessary,
we actually just need to wait until the snapshot is taken before testing
that the exported position is as expected.

(cherry picked from commit 5972414ca00896a8209c634a728a791e023d6d0d)