docs(README): alter release lifecycle to c8 plan
Since the release of Camunda Platform 8, and Zeebe 8.0, Zeebe has
started to follow a different release cadence.

Of course this schedule should be documented correctly in our README.

Minor releases now come less often, but they are supported for longer.
The alpha release candence stays the same, but more alphas are released
before a minor version is released.