feat(engine): activate parallel gateway through modification
By keeping track of the ancestor flow scopes of the element we can increment the number of taken sequence flows in the ProcessInstanceModifiedEventApplier.
Here we have a small assumption that the highest key in the ancestor keys is the direct flowscope of the element that should be activated.

(cherry picked from commit ecdb15de5c11b87a885077ae96e450991b75c1b4)