refactor(engine): deprecate property whilst ensuring backwards compatibility
The activatedElementInstanceKeys property has been moved from the record to the activate instructions. Due to potential replaying of old events we cannot remove the property entirely. Instead the property has been deprecated.

Since we don't populate this property anymore we have to make sure we can still access all the keys. During a replay of an old event we need to access them from the deprecated property, otherwise we need to access them from the activate instructions. By modifying the getter to combine the two behavior will stay the same, regardless of which property the keys come from.