refactor(test-util): retry prematurely closed channel
In a few instances, the Netty channel may be closed prematurely from the
client side - for example, if a read ends up reading less data than
expected, the channel will be silently closed by Netty. In most cases,
it's perfectly safe to retry, as if the server was unavailable we would
get a connection refused error instead. Thus we add a basic retry
mechanism, up to 3 times, in case the channel is closed implicitly
before the SSL handshake could take place.