feat(engine): reset active sequence flows when  terminate end event is completed
After a terminate end event has been completed the flow scope should be completed as well. When there are active sequence flows this doesn't happen as we have checks preventing a flow scope from completing when there is any active element instance or active sequence flow. By resetting this when the terminate end event is completed we can pass this check.