feat(broker): set the default backpressure to s-aimd
S-aimd is more intuitive to configure and performs as good as the default one. By configuring it correctly, we can also increase the limit, thus reducing the backpressure in normal workloads. This might have an impact on overall performance behaviour of the system. In our benchmark setup, the s-aimd had better throughput but slightly higher process execution latency.