ci(.github): setup benchmark workflow
Introduces a benchmark workflow that can create and start the weekly
medic benchmark.

Creating and starting this medic benchmark was always a tedious task.
Luckily most of it was scripted away, so all we have to do is run that
script locally. However, then you have to hope it works on your machine
and wait for everything to complete before you can really continue with
other tasks.

By moving this work into a GitHub action workflow, we can make this
process more reliable (it will only run on a GitHub runner that is
pretty much a clean-slate). And we remove the human step to run the
script and wait for it to finish.

This commit only provides a first try for this workflow and does not yet
push any commits to the medic-cw-benchmark branch, nor does it push the
created docker images. It also doesn't yet build the correct docker
image, which requires additional build-args. Lastly, it does not yet
build the benchmark images.