fix(backup): don't try to delete non-deletable data folder
The data folder, when running inside a container, is mounted as a volume
and is not delete-able. This would cause logging noise and confusion
when restoring would fail. With this change, we only delete the folder
contents and not the folder itself.

(cherry picked from commit 891722f5051d346669c2cf654155ed8410060daa)