test(dispatcher): fixes sigsev due to consumer
While we close the dispatcher last after the actors, one underlying
problem is that the subscription lives "outside" of the actor, and is
not closed when the actor is closed (as it lives in the dispatcher). So
when new fragments are available, the subscription is woken up even if
the consuming actor is closed. This can result in the subscription being
woken up after the actor is closed, after the dispatcher is closed, but
before the scheduler is closed, resulting in a SIGSEV due to accessing
the freed buffer.

By wrapping the polling in an actor call, we can guarantee that it will
only poll if the consuming actor is not closed.

(cherry picked from commit a11119e72bc5de9705f161b07d2dbba2ab4b82bf)