test(engine): unflake timerstartevent tests
Pretty much all timer start event tests increase the engine time to test
the timer scheduling. These tests can all suffer from unwanted race
conditions becaue the timer is scheduled in a post-commit task. There is
no clear way to await for this post-commit task to be executed, except
for awaiting the engine to reach the end of the log.

Since we generally only want to time travel in tests when there is
something related to the timer scheduling, it makes sense to hide this
in the `increaseTime` method of the EngineRule.

Also note that this only matters when the streamprocessor is currently
processing. If the engine is paused, the hasReachedEnd method cannot be
called because the relevant actor is also paused.

Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>
Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>