test(engine): pause engine instead of stopping it
If you want to increase the time, we're now using the streamprocessor
directly. But in the shouldAvoidTriggeringMultipleTimes test case, the
engine was stopped before increasing the time. When the engine is
stopped, the stream processor is removed.

However, this test just wants to verify wht happens if the scheduled
timer triggers and the upcoming due date would also be in the past. That
can be achieved as well by pausing the engine instead of stopping it
completely.

Note that when the engine is paused, we don't have to reset the
recording exporter (on restart it would re-export all records again).

Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>
Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>
(cherry picked from commit c88c9cbfca81de3117dbf6672bb50cded09bba6f)