test(qa): increase timeout of complete topology
Increases the timeout we wait for the topology to be complete from 15
seconds to 30 seconds. The idea is that 15 may be right on the cusp of
the minimum limit we've observed, due to possibly requiring up to 5
seconds for a node to be elected, and hypothetical long GC pauses. This
doesn't not unflake the test, but hopefully the next time it fails we
have a better idea of what is happening.

At the same time, the test is updated to junit 5, and the gateway (from
which we get the topology) is also configured for faster gossip, like
the broker.