test(engine): unflake TimerCatchEventTest
This test was flawed as it creates a timer with a duration of 100ms. Then it continues to verify it has been triggered after 1 second.

This worked because the engine increased it's time. This happened before the timer was scheduled so in reality the timer would trigger after 1.1 seconds. With the change in the increase time method to wait until the engine is idle before increasing the time this broke, as the timer would now be scheduled for the 100ms mark as expected.

(cherry picked from commit 5c467e439fefd179cd86aa0ad8112f1a059a3305)