ci(docker): add empty verify.sh script
This script is called during the Jenkins pipeline for releases (as part
of the Docker image pipeline, see .ci/pipelines/docker_zeebe.groovy).

The script would not be called if the verification is skipped using a
param, however, this param is not set correctly on v8.0. It is set in
release_zeebe.groovy, but on the stable/8.0 branch, which is not used to
run the actual pipeline when releasing 8.0 versions. See
https://github.com/camunda/zeebe/blob/stable/8.0/.ci/pipelines/release_zeebe.groovy#L156

We can simply allow calling the verification script by adding it, but
keeping it empty.

For completeness, I've added a simple description of the script. If we
ever want to add stuff to it, then that's still possible. We just need
to make sure to return the correct exit code.