feat(backup/s3): compress files during upload
If enabled, files larger than threshold (currently 1MiB) are compressed
using Zstandard.
Before uploading, a temporary, compressed copy of the file is created.

While it might be more efficient to compress and archive multiple files
at once, the archive would be more likely to approach 5GiB which is the
limit as long as we are not using multipart upload.