test(engine): remove MultiPartitionDeploymentRecoveryTest
This test was flaky. After analyzing why this is we came to the conclusion that this test is currently not testing what is intended, which would be verifying that a deployment works after recovering from a failed deployment partition.

We still want a test for this, but it proves difficult to this in the engine tests. Instead, we will create a new integration which takes care of this.

(cherry picked from commit b3f336178334fdd5a67bd7a5444b6f4ba5c5f950)