test(qa): verify deployment redistribution on partition restart
This test verifies that a deployment gets redistributed in the event of a deployment partition restart. In order to achieve this we pause one of the partition of the cluster. This is to make sure we don't process the deployment before we stop the deployment partition.

Once we've verified that the 3rd partition has finished it's deployment as expected we will resume the second partition and restart the deployment partition. After this the test verifies that the deployment is completed as expected.

(cherry picked from commit f43905ec1841930c53a4ee0dfb2b0c96990d50de)