feat(backup/s3): add metadata on snapshot and segment files to manifest
This updates all manifest types to use `FileSet` instead of a plain
list of file names. When uploading compressed snapshot or segment files,
the manifest is updated to include the used compression algorithm.

Backwards compatability is maintained by adding aliases to the key names
in the manifest and by using the custom deserializer of `FileSet`.