feat(exporter/elasticsearch): set the record sequence
Set the sequence number when adding a record to the bulk request.

The sequence is not part of the record interface. To serialize the sequence together with the record, we need to deserialize the record first as a map, add the sequence, and finally serialize the map into a JSON.

The additional serialization and deserialization adds more load to the Elasticsearch exporter. But we don't pollute the record interface with a sequence that is only required by the Elasticsearch exporter.