refactor: split engine module
Split engine module into two separate modules.

One called stream-platform which contains everything related to stream
processing. This module contains right now the API and implementation.
It can and will be splitted further in the future.

The other remaining module is still called engine, which contains the
record processor implementation. It is the core or business logic of
zeebe (the process engine).

Implementation classes have been moved and imports are adjusted in this
commit.

There are several classes which needed to be moved, which we might want
to move to a different place later.

For example:

 * ZbColumnFamily
 * CopiedRecords
 * RecordValues
 * etc.