feat(exporter/elasticsearch): manage the record counters
Pass a sequence number when indexing a record. The sequence is created based on the partition id of the record and a counter.

The exporter manages the counter. It increments the counter for each exported record.

The counter is grouped by value type. Records with the same value type use the same counter. Other records with different value types use other counters. As a result, the sequence of all records with one value type is increasing and there is no gap between two records.

Since the records are stored in Elasticsearch indexes by their value type, a consumer can use the sequence to read a bulk of records based on its last read sequence and the bulk size.