ci(.github): dont overwrite cache for java-client
The `java-client` job first builds the project using the required jdk
and then switches the jdk when testing that the client works on java 8.

This `cache: maven` line would overwrite the cache, incl. all build
modules. So the test did not run on the correct code.

By removing the line, we avoid that the modules compiled in the
build-zeebe step are not overwritten.