fix(engine): prevent NPE when terminating call activity
If a parent process has an active call activity element instance there is no guarantee that the called activity is still activate. When the called activity is completed it will try to complete the call activity element. At this point we try to apply the output mappings. If something goes wrong here an incident is created.
When we try to terminate the call activity at this point the called activity is already completed and no instance of it will be available. For this reason we must verify that this instance is not null before we try to terminate it.

(cherry picked from commit 254a863703d540b365e36ca63ac4995d2bd2d343)