test(exporter): don't fail test if `getPhase` future throws
Previously, we added an await on the exporter phase to prevent race
conditions in this test that would result in a flaky test.

As it turns out, getting the exporter phase is itself also racy and
might return a future that is completed exceptionally because the actor
got closed between calling `getPhase` and before executing the actor
call.

As a fix, we simply ignore exceptions while waiting for a closed
exporter.

(cherry picked from commit d3a89b13011e7e09f7e70dac53439ef6f65bce7c)