feat(backups/s3): opportunistically compress files before uploading
If backup compression is enabled and the fileset manager has to upload a
file that is larger than a constant threshold (currently 8MiB, chosen
arbitrarily), it first writes zstd compressed content to a temporary
file and uploads that file instead. The file metadata is set so that
the fileset manager knows what compression algorithm (if any) to use
when restoring.