fix(test): resolve flaky test
This test was flaky because the margin of the between was too tight. The CI isn't always fast enough to ensure the timer gets triggered between 100-150ms.

Since we don't really like timers of 100ms we changed to a more realistic scenario with a timer of 1 minute. We also increased the margin to have an extra minute to prevent flakyness