refactor(test): rename changesFlowScope to force
An execution path segment does not know about flow scopes and processes. All it knows is that there is specific steps that need to be executed. It stores a list of steps.

It doesn't make sense for this parameter to be named changesFlowScope. This class doesn't care about it, it cares about whether it should append a segment to another segment. Therefore, force is a better name for this parameter.