build(.github): remove automated version computation
Hardcodes the stable versions into the workflow. The main reason here is
because we would inevitably get into problems when a new release is
done, since typically the generation doesn't exist yet at that point,
which would cause our workflow to fail.

Once we figure out some way to guarantee the generation also exists,
or a way to only pick versions which have a generation, then we can move
forward with this.