feat(qa-util): add ProxyRegistry utility
Adds a new `ProxyRegistry` utility, since Testcontainers deprecated this
functionality from the main library.

The registry keeps track of all proxies created with a given Toxiproxy
container, which is quite useful to be able to link cluster nodes with
their proxies in a non-blocking way. The registry is designed based on
its usage right now, so should be evolved naturally if that changes.