feat: use selected ancestor for activation
This adds the capability to the ElementActivationBehavior (used by both
the process instance modification, as well as the create instance with
start instructions), to select the ancestor of the element to activate.

This is necessary when there are multiple instances of the flow scope of
the element to activate, and so the engine would not be able to choose
the right flow scope for the newly activated element.

For process instance creation with start instructions, the ancestor
selection is disabled because there is no need for ancestor selection.
There is no ancestor to select because the instance doesn't even exist
yet. It is going to be created.