feat(backup/s3): allow configuration of the compression algorithm
This changes the boolean flag `enableCompression` to a string value
that is either 'none' which disables compression or the name of a
supported compression algorithm.

Because the configuration class was getting a bit unwieldy, we are also
adding a builder class that allows for more ergonomic configuration.

The config constructor now validates configuration, in particular to
check that the bucket name is not empty and that the named algorithm is
supported for both compression and decompression