feat: log details of cancelled transitions
When one transition cancels another, we were only logging a generic
"cancel" message. This was lacking context, specifically which
transition causes a cancellation.

Here we add a `toString` implementation for `PartitionTransitionProcess`
that shows term, role, status and the remaining steps. This is used in
a new log statement that should provide us with the necessary context.